Sentiment_Analysis/Financial-News-Analysis/requirements.txt
pandas
numpy
nltk
scikit-learn
joblib
streamlit
sentence_transformers
imblearn
imbalanced-learn
google-generativeai
dotenv
